test: update silent audio regression baselines
