+++ packages/producer/tests/font-variant-numeric/output/output.mp4
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d2e098b9bd60796b15b6945a75bf4b7e9bd0939f806e0501ef59922d9522d0a6
-size 142963
+oid sha256:506ae124493ffc9bfc543eae709106249646069937014ded2d6edbe661ce832a
+size 151013
