+++ packages/producer/tests/missing-host-comp-id/output/output.mp4
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7a2e53926e44d66a469e74eee88035987e309786eb44fef17de9f34114de0813
-size 105638
+oid sha256:a84e50ad31694432affb1bffd42bde653514662ccb1dbb8172b0eab2d80147d5
+size 129943
